# public/swagger-stock-api.json
{
    "swagger": "2.0",
    "info": {
        "title": "Stock Monitor API",
        "description": "Stock Monitoring Service",
        "version": "1.0.0"
    },
    "host": "localhost:3000",
    "schemes": [
        "https"
    ],
    "basePath": "/recipes",
    "produces": [
        "application/json"
    ],
    "paths": {
			
		"/stockTrigger": {
		      "post": {
              "summary": "Add a new stock recipe",
              "description": "This endpoint is used to add a new recipe with a stock trigger for the stock monitoring service",
                "parameters": [
                    {
                        "name": "Recipe with Stock Trigger",
                        "in": "body",
                        "description": " JSON to be created",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/newSRecipe"
                        }
                    }
                ],
                "tags": [
                    "stock"
                ],
                "responses": {
                    "200": {
                        "description": "Success message",
                        "schema": {
                            "$ref": "#/definitions/Success"
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },

        "/recipes/:recipeid": {
            "delete": {
              "summary": "Delete a specific recipe",
              "description": "This endpoint is used to delete any recipe for this service",
                "parameters": [
                    {
                        "name": "recipeID",
                        "description": " Unique recipe identifier",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Delete"
                        }
                    }
                ],
                "tags": [
                    "Recipes"
                ],
                "responses": {
                    "200": {
                        "description": "Success message",
                        "schema": {
                            "$ref": "#/definitions/Success"
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "Delete": {
            "properties": {
                "recipeID": {
                    "type": "string",
                    "description": "Unique identifier representing a specific" 
                }
            }
        },
        "sTrigger": {
            "properties": {
                "watchNum": {
                    "type": "integer",
                    "description": "Value to watch for if the relations calls for it."
                },
                "symbol": {
                    "type": "string",
                    "description": "Symbol of stock.",
					"required": true
                },
				"market": {
                    "type": "string",
                    "description": "The market the stock is in.",
					"required": true
                },
                "relation": {
                    "type": "string",
                    "description": "Type of stock trigger (stockGT, stockLT, stockPerInc, stockPerDec, closePrice)",
					"required": true
                },
				"inThreshold": {
					"type": "boolean",
					"description": "Always set to false. Used for limiting the amount of times a recipe can trigger."
				}
            }
        },
		
		"newSRecipe": {
			"properties": {
				"recipe": {
				    "required": true,
                    "description": "List of trigger specifications",
                    "$ref": "#/definitions/sRecipe"
				}
			}
		},
	    "sRecipe": {
            "properties": {
                "callbackURL": {
                    "type": "string",
                    "description": "Callback URL to call when the trigger gets hit."
                },
                "trigger": {
                    "description": "List of trigger specifications",
                    "$ref": "#/definitions/sTrigger"
                }
            }
        },		
        "Success": {
            "properties": {
                "success": {
                    "type": "boolean",
                    "description": "Whether the request was successful or not"
                },
                "msg": {
                    "type": "string",
                    "description": "Success or error message"
                },
                "recipe": {
                    "description": "Recipe with weather trigger object returned",
                    "$ref": "#/definitions/newRecipe"
                }
            }
        },
        "Error": {
            "properties": {
                "success": {
                    "type": "boolean"
                },
                "msg": {
                    "type": "string"
                }
            }
        }
    }
}